+++ terrain/voxel_lod_terrain.h
@@ -211,6 +211,10 @@ public:
 	bool is_showing_gizmos() const {
 		return _show_gizmos_enabled;
 	}
+	void set_show_octree_gizmos(bool enable);
+	bool is_showing_octree_gizmos() const {
+		return _show_octree_node_gizmos;
+	}
 #endif
 
 	TypedArray<String> get_configuration_warnings() const override;